// ui/src/lib/gen/TerminalPalette.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

/**
 * Colours used to answer OSC 4/10/11/12 queries (`#rrggbb`).
 */
export type TerminalPalette = { foreground: string, background: string, cursor: string, 
/**
 * 16 ANSI colours (0-15).
 */
ansi: Array<string>, };